feat(teamProjectsHeader): Reorganized file locations
