+++ src/Analysis/SharedUtils/TeamProject/Utils/CheckForTeamProjectApplication.js
